add eventhubmetricsprovidertests, port over tests from scale monitor and remove unnecessary imports
